+++ Cargo.toml
@@ -1,27 +1,43 @@
 [package]
 name = "ruspiro-register"
 authors = ["André Borrmann <pspwizard@gmx.de>"]
-version = "0.4.3" # remember to update html_root_url
+version = "0.5.0" # remember to update html_root_url in lib.rs
 description = """
-The crate provides simple to use macros to define MMIO registers representations allowing safe access to their contents.
-In addition this crate defines specific aarch64 and aarch32(CP15) registers to safely use them in rust code. This
-reduces the need of 'unsafe' calls.
+The crate provides the definitions to conviniently work with register field values that are typically presented by a set of bit fields.
 """
-license = "Apache-2.0"
-repository = "https://github.com/RusPiRo/ruspiro-register/tree/v0.4.3"
-documentation = "https://docs.rs/ruspiro-register/0.4.3"
+license = "MIT OR Apache-2.0"
 readme = "README.md"
-keywords = ["RusPiRo", "register", "raspberrypi", "baremetal", "mmio"]
+repository = "https://github.com/RusPiRo/ruspiro-register/tree/v||VERSION||"
+documentation = "https://docs.rs/ruspiro-register/||VERSION||"
 categories = ["no-std", "embedded"]
+keywords = ["ruspiro", "register", "registerfields"]
 edition = "2018"
 exclude = [".travis.yml", "Makefile.toml"]
 
 [badges]
-travis-ci = { repository = "RusPiRo/ruspiro-register", branch = "master" }
+travis-ci = { repository = "RusPiRo/ruspiro-register", branch = "release" }
 maintenance = { status = "actively-developed" }
+is-it-maintained-open-issues = { repository = "RusPiRo/ruspiro-register" }
 
 [lib]
 
+[build-dependencies]
+# uncomment this if a build.rs script should be run as part of the build process
+# cc = "1.0"
+
 [dependencies]
+# list other dependencies here
+# for every RUSPIRO dependency a corresponding entry
+# in the [patch.crates-io] section pointing to the actual master branch
+# of this crate is required - example:
+# ruspiro-register = "0.4"
 
 [features]
+ruspiro_pi3 = [ ]
+
+[patch.crates-io]
+# we require an entry for each dependent RUSPIRO crate here
+# this increases PR build stability if dependent crates got updated but are not
+# published to crates.io yet.
+# !This section need to be always the last in the Cargo.toml file!
+# ruspiro-register = { git = "https://github.com/RusPiRo/ruspiro-register.git" }
