+++ Makefile.toml
@@ -1,40 +1,45 @@
 #***********************************************************************************************************************
 # cargo make tasks to build the example for the Raspberry Pi
 #***********************************************************************************************************************
-[env.a64]
+
+# AARCH64 specific profile environment varialbles
+[env.development]
 CC = "aarch64-none-elf-gcc"
 AR = "aarch64-none-elf-ar"
-OC = "aarch64-none-elf-objcopy"
 CFLAGS = "-march=armv8-a -Wall -O3 -nostdlib -nostartfiles -ffreestanding -mtune=cortex-a53"
 RUSTFLAGS = "-C linker=${CC} -C target-cpu=cortex-a53 -C link-arg=-nostartfiles -C link-arg=-T./link64.ld"
-BUILD_TARGET = "aarch64-unknown-linux-gnu"
+BUILD_TARGET = "aarch64-unknown-none"
 
-[env.a32]
-CC = "arm-none-eabi-gcc"
-AR = "arm-none-eabi-ar"
-OC = "arm-none-eabi-objcopy"
-CFLAGS = "-mcpu=cortex-a53 -march=armv7-a -mfpu=neon -mfloat-abi=softfp -Wall -O3 -nostdlib -nostartfiles"
-RUSTFLAGS = "-C linker=${CC} -C target-cpu=cortex-a53 -C link-arg=-nostartfiles"
-BUILD_TARGET = "armv7a-none-eabi"
-
-# Travis CI need a different CC/AR alltogether as the aarch64-none-elf is not available there as it seems
-[env.a64-travis]
+# AARCH64 specific Travis CI env. variables. "aarch64-none-elf" is not available there as it seems
+[env.travis]
 CC = "aarch64-linux-gnu-gcc"
 AR = "aarch64-linux-gnu-ar"
-OC = "aarch64-linux-gnu-objcopy"
 CFLAGS = "-march=armv8-a -Wall -O3 -nostdlib -nostartfiles -ffreestanding -mtune=cortex-a53"
-RUSTFLAGS = "-C linker=${CC} -C target-cpu=cortex-a53 -C link-arg=-nostartfiles"
-BUILD_TARGET = "aarch64-unknown-linux-gnu"
+RUSTFLAGS = "-C linker=${CC} -C target-cpu=cortex-a53 -C link-arg=-nostartfiles -C link-arg=-T./link64.ld"
+BUILD_TARGET = "aarch64-unknown-none"
 
 [tasks.xbuild]
 command = "cargo"
 args = ["xbuild", "--target", "${BUILD_TARGET}", "--release", "--features", "${FEATURES}"]
 
 [tasks.pi3]
-env = { FEATURES = "" }
-run_task = [
-    { name = "xbuild" }
-]
+env = { FEATURES = "ruspiro_pi3" }
+run_task = "xbuild"
+
+[tasks.clippy]
+env = { FEATURES = "ruspiro_pi3" }
+command = "cargo"
+args = ["clippy", "--target", "${BUILD_TARGET}", "--features", "${FEATURES}"]
+
+[tasks.doc]
+env = { FEATURES = "ruspiro_pi3" }
+command = "cargo"
+args = ["doc", "--target", "${BUILD_TARGET}", "--open"]
+
+[tasks.doctest]
+env = { FEATURES = "ruspiro_pi3" }
+command = "cargo"
+args = ["xtest", "--doc", "--target", "${BUILD_TARGET}"]
 
 [tasks.clean]
 command = "cargo"
