--- src/system/aarch32/mod.rs
@@ -1,16 +1,0 @@
-/***********************************************************************************************************************
- * Copyright (c) 2019 by the authors
- *
- * Author: André Borrmann
- * License: Apache License 2.0
- **********************************************************************************************************************/
-
-//! # Aarch32 CP15 System Register
-//! Aarch32 register definitions are only available when compiled for **Arm** target architecture
-
-pub mod actlr;
-pub mod dacr;
-pub mod sctlr;
-pub mod ttbcr;
-pub mod ttbr0;
-pub mod ttbr1;
